% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2shp.r
\name{gl2shp}
\alias{gl2shp}
\title{Converts a genlight object to ESRI shapefiles or kml files}
\usage{
gl2shp(
  x,
  type = "shp",
  proj4 = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  outfile = "gl",
  outpath = tempdir(),
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data and location
data, lat longs [required].}

\item{type}{Type of output 'kml' or 'shp' [default 'shp'].}

\item{proj4}{Proj4string of data set (see spatialreference.org for
projections) [default WGS84].}

\item{outfile}{Name (path) of the output shape file [default 'gl']. shp
extension is added automatically.}

\item{outpath}{Path where to save the output file; the folder must exist
[default tempdir(), mandated by CRAN]. Use outpath=getwd() or outpath='.'
when calling this function to direct output files to your working directory.}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
brief progress messages; 3, progress and results summary; 5, full report
[default 2, unless specified using gl.set.verbosity].}
}
\value{
A terra SpatVector with the exported points. The file is written to
outpath.
}
\description{
This function exports the coordinates in a genlight object to a point shape
file or kml file, with the individual metrics as attributes.
}
\details{
Coordinates are taken from x@other$latlon (columns lat and lon; a column
named long is accepted for lon) and are assumed to be in WGS84 unless a
proj4 string is provided.

Individuals with a missing lat or lon are removed. Missing values in other
individual metrics are kept and written as NA.

The attribute table holds the columns of x@other$ind.metrics plus lat and
lon. If ind.metrics has no id column, an id column with the individual
names is added. ESRI shapefiles truncate column names to 10 characters.
}
\examples{
out <- gl2shp(testset.gl, outpath=tempdir())

}
\author{
Author(s): Bernd Gruber. Custodian: Bernd Gruber -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
\concept{linker}
