% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{zzz}
\alias{zzz}
\title{Setting up the package dartR.popgenomics}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
zzz
}
\description{
Setting up dartR.popgenomics
}
\keyword{internal}
