% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.grm2.r
\name{gl.grm2}
\alias{gl.grm2}
\title{Calculates an identity by descent matrix}
\usage{
gl.grm2(
  x,
  plotheatmap = TRUE,
  palette_discrete = NULL,
  palette_convergent = NULL,
  legendx = 0,
  legendy = 0.5,
  label.size = 0.75,
  legend.title = "Populations",
  plot.file = NULL,
  plot.dir = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{plotheatmap}{A switch if a heatmap should be shown [default TRUE].}

\item{palette_discrete}{the color of populations [gl.select.colors].}

\item{palette_convergent}{A convergent palette for the IBD values
or vector of colours [default RdYlBu from gl.select.colors].}

\item{legendx}{x coordinates for the legend[default 0].}

\item{legendy}{y coordinates for the legend[default 0.5].}

\item{label.size}{Specify the size of the population labels [default 0.75].}

\item{legend.title}{Legend title [default "Populations"].}

\item{plot.file}{Name for the RDS binary file to save (base name only,
exclude extension); saved only when \code{plotheatmap = TRUE}
[default NULL]}

\item{plot.dir}{Directory in which to save files [default = working directory]}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log ; 3, progress and results summary; 5, full report
[default 2 or as specified using gl.set.verbosity].}

\item{...}{Parameters passed to function A.mat from package rrBLUP.}
}
\value{
An identity by descent matrix
}
\description{
This function is the dartR.spatial copy of \code{gl.grm} (package
dartR.captive), kept so that spatial analyses do not depend on
dartR.captive. Both return the same matrix.

This function calculates the mean probability of identity by state (IBS)
across loci that would result from all the possible crosses of the
individuals analyzed. IBD is calculated by an additive relationship matrix
approach developed by Endelman and Jannink (2012) as implemented in the
function \link[rrBLUP]{A.mat} (package rrBLUP).
}
\details{
This function uses the A.mat function from the rrBLUP package. This method 
follows the approach developed by Endelman and Jannink (2012).

Two alleles are Identical by State (IBS) if they are the same in state, 
regardless of whether they come from a common ancestor. Two alleles are 
Identical by Descent (IBD) if they are inherited from a common ancestor. 
While IBS does not necessarily imply IBD, using high-density SNP data 
improves the estimation of IBD probabilities from IBS measures.

This function also plots a heatmap, and a dendrogram, of IBD values where
each diagonal element has a mean that equals 1+f, where f is the inbreeding
coefficient (i.e. the probability that the two alleles at a randomly chosen
locus are IBD from the base population). As this probability lies between 0
 and 1, the diagonal elements range from 1 to 2. Because the inbreeding
 coefficients are expressed relative to the current population, the mean of
 the off-diagonal elements is -(1+f)/n, where n is the number of loci.
 Individual names are shown in the margins of the heatmap and colors
 represent different populations.

 This function densifies the genotype matrix (\code{as.matrix(x)}) before
 calling \link[rrBLUP]{A.mat}, which requires a dense matrix. It is not
 suited to full-size FBM-backed objects, where densification defeats the
 memory savings FBM backing is meant to provide.
}
\examples{
if (isTRUE(getOption("dartR_fbm"))) platypus.gl <- gl.gen2fbm(platypus.gl)
gl.grm2(platypus.gl[1:10, 1:100])

}
\references{
\itemize{
\item Endelman, J. B. (2011). Ridge regression and other kernels for genomic
 selection with r package rrblup. The Plant Genome 4, 250.
\item Endelman, J. B. , Jannink, J.-L. (2012). Shrinkage estimation of the
realized relationship matrix. G3: Genes, Genomics, Genetics 2, 1405.
}
}
\seealso{
\code{gl.grm} and \code{gl.grm.network} in package dartR.captive,
which compute the same matrix.
}
\author{
Author(s): Arthur Georges. Custodian: Arthur Georges -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
\concept{inbreeding functions}
