% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.run.eems.r
\name{gl.run.eems}
\alias{gl.run.eems}
\title{Runs the EEMS algorithm  (Estimating Effective Migration Surfaces) on a genlight object.}
\usage{
gl.run.eems(
  x,
  eems.path = "./",
  buffer = 10000,
  nDemes = 500,
  diploid = TRUE,
  numMCMCIter = 10000,
  numBurnIter = 2000,
  numThinIter = 9,
  seed = NULL,
  out.dir = NULL,
  plot.dir = NULL,
  plot.file = NULL,
  verbose = NULL,
  cleanup = TRUE,
  ...
)
}
\arguments{
\item{x}{Genlight object [required].}

\item{plot.dir}{Directory in which to save output and intermediate files [default = tempdir()]}

\item{plot.file}{Name for the RDS binary file to save the list of plots (base name only, exclude extension) [default eems]}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log ; 3, progress and results summary; 5, full report
[default 2 or as specified using gl.set.verbosity].}
}
\value{
A list of contour plots of migration and diversity rates as well as 
several diagnostic plots It is a good idea to examine all these figures, 
which is why they are generated by default. Please check the examples how to 
customize the figrues
\describe{
 \item{mrates01}{effective migration surface. This contour plot visualizes the estimated effective migration rates \code{m}, on the log10 scale after mean centering.}
 \item{mrates02}{posterior probability contours \code{P(log(m) > 0) = p} and \code{P(log(m) < 0) = p} for the given probability level \code{p}. Since migration rates are visualized on the log10 scale after mean centering, 0 corresponds to the overall mean migration rate. This contour plot emphasizes regions with effective migration that is significantly higher/lower than the overall average.}
 \item{qrates01}{effective diversity surface. This contour plot visualizes the estimated effective diversity rates \code{q}, on the log10 scale after mean centering.}
 \item{qrates02}{posterior probability contours \code{P(log(q) > 0) = p} and \code{P(log(q) < 0) = p}. Similar to \code{mrates02} but applied to the effective diversity rates.}
 \item{rdist01}{scatter plot of the observed vs the fitted between-deme component of genetic dissimilarity, where one point represents a pair of sampled demes.}
 \item{rdist01}{scatter plot of the observed vs the fitted within-deme component of genetic dissimilarity, where one point represents a sampled deme.}
 \item{rdist03}{scatter plot of observed genetic dissimilarities between demes vs observed geographic distances between demes.}
 \item{pilogl01}{posterior probability trace}
}
}
\description{
This function runs the EEMS algorithm on a genlight object. The EEMS algorithm
is a spatially explicit model that estimates effective migration surfaces
(EEMS) from genetic data. The EEMS algorithm is implemented in C++, hence it 
is necessary to have the binary downloaded and the function need to point to 
this file via the path specified in eems.path. The binary is call runeems_snps[.exe] 
and can be downloaded from the github site of dartR.
}
\examples{
 \dontrun{
 #this examples needs a binary (runeems_snps[.exe]) specific to your operating 
 #system  to run
 eems <- gl.run.eems(bandicoot.gl, eems.path = "d:/downloads/eems/")
 print(eems[[1]])
 }

}
\references{
Petkova D (2024). _reemsplots2: Generate plots to inspect and visualize the results
of EEMS_. R package version 0.1.0, <https://github.com/dipetkov/eems>.
D Petkova, J Novembre, M Stephens. Visualizing spatial population structure
with estimated effective migration surfaces. Nature Genetics 48, 94 -- 100 (2016). 
\url{http://dx.doi.org/10.1038/ng.3464}.
}
\author{
Bernd Gruber  & Robyn (bugs? Post to 
\url{https://groups.google.com/d/forum/dartr})
}
